feat: wrap strategies between centralized and decentralized
