+++ substrafl/strategies/__init__.py
@@ -1,7 +1,7 @@
-from substrafl.strategies.fed_avg import FedAvg
-from substrafl.strategies.newton_raphson import NewtonRaphson
-from substrafl.strategies.scaffold import Scaffold
-from substrafl.strategies.single_organization import SingleOrganization
+from substrafl.strategies.centralized_strategies.fed_avg import FedAvg
+from substrafl.strategies.centralized_strategies.newton_raphson import NewtonRaphson
+from substrafl.strategies.centralized_strategies.scaffold import Scaffold
+from substrafl.strategies.decentralized_strategies.single_organization import SingleOrganization
 from substrafl.strategies.strategy import Strategy
 
 __all__ = ["Strategy", "FedAvg", "SingleOrganization", "Scaffold", "NewtonRaphson"]
