+++ tests/algorithms/pytorch/test_base_algo.py
@@ -73,6 +73,9 @@ def rng_strategy():
         def name(self) -> StrategyName:
             return "rng_strategy"
 
+        def initialization_round(*args, **kwargs):
+            pass
+
         def perform_round(
             self,
             algo,
@@ -84,9 +87,6 @@ def rng_strategy():
         ):
             next_local_states = []
             next_shared_states = []
-
-            if round_idx == 0:
-                return
 
             for i, node in enumerate(train_data_nodes):
                 next_local_state, next_shared_state = node.update_states(
